+++ Sibusten.Philomena.Client/DownloadProgressInfo.cs
@@ -1,0 +1,9 @@
+namespace Sibusten.Philomena.Client
+{
+    public record DownloadProgressInfo
+    {
+        public string Action { get; init; } = "";
+        public long Current { get; init; }
+        public long? Total { get; init; }
+    }
+}