+++ Sibusten.Philomena.Client/Extensions/PhilomenaImageSearchExtensions.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Sibusten.Philomena.Client.Images;
 using Sibusten.Philomena.Client.Options;
 
@@ -6,14 +8,24 @@ namespace Sibusten.Philomena.Client.Extensions
     public static class PhilomenaImageSearchExtensions
     {
         /// <summary>
-        /// Gets a downloader for the image search
+        /// Downloads images from a search in parallel
         /// </summary>
         /// <param name="search">The image search</param>
         /// <param name="options">Download options</param>
-        /// <returns>A downloader for the image search</returns>
-        public static IPhilomenaImageDownloader GetDownloader(this IPhilomenaImageSearch search, ImageDownloadOptions? options = null)
+        public static async Task DownloadAllParallel(this IPhilomenaImageSearch search, ParallelPhilomenaImageDownloaderOptions? options = null)
         {
-            return new PhilomenaImageDownloader(search, options);
+            await search.BeginSearch().DownloadAllParallel(options);
+        }
+
+        /// <summary>
+        /// Downloads images in parallel
+        /// </summary>
+        /// <param name="images">The images to download</param>
+        /// <param name="options">Download options</param>
+        public static async Task DownloadAllParallel(this IAsyncEnumerable<IPhilomenaImage> images, ParallelPhilomenaImageDownloaderOptions? options = null)
+        {
+            ParallelPhilomenaImageDownloader downloader = new ParallelPhilomenaImageDownloader(options);
+            await downloader.Download(images);
         }
     }
 }