+++ Sibusten.Philomena.Client/Images/IPhilomenaImageSearch.cs
@@ -4,14 +4,20 @@ using System.Threading;
 
 namespace Sibusten.Philomena.Client.Images
 {
-    public struct MetadataDownloadProgressInfo
+    public record MetadataDownloadProgressInfo
     {
-        public int Downloaded { get; set; }
-        public int Total { get; set; }
+        public int Downloaded { get; init; }
+        public int Total { get; init; }
+    }
+
+    public record ImageSearchProgressInfo
+    {
+        public int Processed { get; init; }
+        public int Total { get; init; }
     }
 
     public interface IPhilomenaImageSearch
     {
-        IAsyncEnumerable<IPhilomenaImage> BeginSearch(CancellationToken cancellationToken = default, IProgress<MetadataDownloadProgressInfo>? progress = null);
+        IAsyncEnumerable<IPhilomenaImage> BeginSearch(CancellationToken cancellationToken = default, IProgress<ImageSearchProgressInfo>? searchProgress = null, IProgress<MetadataDownloadProgressInfo>? metadataProgress = null);
     }
 }