+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 
 ### Changed
 - Split up `PhilomenaImageSearchQuery` to decouple logic
-    - Extension method `IPhilomenaImageSearch.GetDownloader` can be used to get a downloader from a search
+    - Add new extension methods to simplify use of the new classes
 - Move image related classes to `Images` namespace
 - Download images to temp files before moving to target files